+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   [#338](https://github.com/phadej/github/pull/338)
 - Allow multiple assignees in NewIssue and EditIssue
   [#336](https://github.com/phadej/github/pull/336)
+- Add `pullRequestPatchR` and `pullRequestDiffR`
 
 ## Changes for 0.20